add challenge id verification
