+++ sdk/viem/src/custom-actions/verify-wallet-verification-challenge.action.ts
@@ -13,11 +13,21 @@ export type AddressOrObject<Extra = {}> =
     } & Extra);
 
 /**
+ * Represents either a wallet address string, an object containing wallet address and optional verification ID or a challenge ID.
+ */
+export type AddressOrObjectWithChallengeId =
+  | AddressOrObject
+  | {
+      /** ID of the challenge to verify against */
+      challengeId: string;
+    };
+
+/**
  * Parameters for verifying a wallet verification challenge.
  */
 export interface VerifyWalletVerificationChallengeParameters {
   /** The wallet address or object containing wallet address and optional verification ID. */
-  addressOrObject: AddressOrObject;
+  addressOrObject: AddressOrObjectWithChallengeId;
   /** The response to the verification challenge. */
   challengeResponse: string;
 }
@@ -40,7 +50,7 @@ export type VerifyWalletVerificationChallengeResponse = VerificationResult[];
  */
 type WalletRpcSchema = {
   Method: "user_verifyWalletVerificationChallenge";
-  Parameters: [addressOrObject: AddressOrObject, challengeResponse: string];
+  Parameters: [addressOrObject: AddressOrObjectWithChallengeId, challengeResponse: string];
   ReturnType: VerifyWalletVerificationChallengeResponse;
 };
